+++ Example/SJMediaCacheServer/Lib/Resource/SJResourceReader.h
@@ -16,8 +16,8 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, weak, nullable) id<SJResourceReaderDelegate> delegate;
 
 - (void)prepare;
-@property (nonatomic, readonly) UInt64 contentLength;
-@property (nonatomic, readonly) UInt64 offset;
+@property (nonatomic, readonly) NSUInteger contentLength;
+@property (nonatomic, readonly) NSUInteger offset;
 - (nullable NSData *)readDataOfLength:(NSUInteger)length;
 @property (nonatomic, readonly) BOOL isReadingEndOfData;
 - (void)close;
